实现角色管理；优化部分页面样式；

+++ manage-service/src/main/java/com/ozf/laiyw/manage/service/impl/MenuServiceImpl.java
@@ -74,7 +74,10 @@ public class MenuServiceImpl implements MenuService {
         List list = menuMapper.getChildrenByParentId(id);
         if (null != list && list.size() > 0)
             return -1;
-        //TODO 该菜单是否有被引用
+        int num = menuMapper.isQuote(id);
+        if (num > 0) {
+            return -2;
+        }
         int count = menuMapper.removeMenu(id);
         initData();
         return count;
